互娱商城购买咖豆下单(预提交)、推广员下属成员购买咖豆回调 --task=1009086 --user=a0303邹荣亮开发工程师 【服务端】新互娱推广系统 https://www.tapd.cn/66077311/s/1290694

+++ pkg/common/db/agent_model_struct.go
@@ -94,7 +94,7 @@ type TAgentBeanRechargeOrder struct {
 	OrderNo       string    `json:"order_no"`        // 平台订单号
 	ChessOrderNo  string    `json:"chess_order_no"`  // 互娱订单号
 	NcountOrderNo string    `json:"ncount_order_no"` // 新生支付订单号
-	Number        int32     `json:"number"`          // 充值数量
+	Number        int64     `json:"number"`          // 充值数量
 	GiveNumber    int32     `json:"give_number"`     // 赠送金额
 	Amount        int32     `json:"amount"`          // 金额(单位:元)
 	PayTime       int32     `json:"pay_time"`        // 支付时间
